+++ FinalCattery/src/main/java/by/epam/cattery/controller/command/impl/admin/DeleteReviewCommand.java
@@ -7,6 +7,7 @@ import by.epam.cattery.controller.command.constant.SessionConst;
 import by.epam.cattery.controller.content.NavigationType;
 import by.epam.cattery.controller.content.RequestContent;
 import by.epam.cattery.controller.content.RequestResult;
+import by.epam.cattery.entity.Role;
 import by.epam.cattery.util.ConfigurationManager;
 import by.epam.cattery.service.ReviewService;
 import by.epam.cattery.service.ServiceFactory;
@@ -19,16 +20,22 @@ public class DeleteReviewCommand implements ActionCommand {
     private static final Logger logger = LogManager.getLogger(DeleteReviewCommand.class);
 
     private static final String SUCCESS_PAGE = ConfigurationManager.getInstance().getProperty(PathConst.SUCCESS_PAGE);
+    private static final String ACCESS_DENIED_PAGE = ConfigurationManager.getInstance().getProperty(PathConst.ACCESS_DENIED_PAGE);
 
 
     @Override
     public RequestResult execute(RequestContent requestContent) throws ServiceException {
-        ReviewService reviewService = ServiceFactory.getInstance().getReviewService();
+        String path = ACCESS_DENIED_PAGE;
 
-        int userId = Integer.parseInt(requestContent.getParameter(SessionConst.ID));
-        int reviewId = Integer.parseInt(requestContent.getParameter(RequestConst.REVIEW_ID));
-        reviewService.deleteReview(reviewId, userId);
+        if (requestContent.getSessionAttribute(SessionConst.ROLE) == Role.ADMIN) {
+            ReviewService reviewService = ServiceFactory.getInstance().getReviewService();
 
-        return new RequestResult(NavigationType.REDIRECT, SUCCESS_PAGE);
+            int userId = Integer.parseInt(requestContent.getParameter(SessionConst.ID));
+            int reviewId = Integer.parseInt(requestContent.getParameter(RequestConst.REVIEW_ID));
+            reviewService.deleteReview(reviewId, userId);
+            path = SUCCESS_PAGE;
+        }
+
+        return new RequestResult(NavigationType.REDIRECT, path);
     }
 }
